automation/api/requirements.txt
bosdyn-client==4.0.1
python-dotenv >= 1.0.0
flask >= 2.2.5
uwsgi >= 2.0.22
flask-cors >= 3.0.10
ping3 >= 4.0.4
